gala-spider/requirements.txt
connexion >= 1.1.15
python_dateutil == 2.6.0
setuptools >= 21.0.0
requests
pyyaml
pyarango == 2.0.1